% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_to_sql.R
\name{compile_to_sql}
\alias{compile_to_sql}
\title{Compile picorule text to SQL}
\usage{
compile_to_sql(picorule_name, picorule_text, target_sql, compiler_api)
}
\arguments{
\item{picorule_name}{Name of the picorule block (AKA block_id)}

\item{picorule_text}{Body of the picorule syntax}

\item{target_sql}{Target SQL framework}

\item{compiler_api}{URI for API}
}
\value{
text with SQL code
}
\description{
Compile picorule text to SQL
}
\examples{
compile_to_sql("ckd",".","PLSQL","http://localhost:8321/v1/engine/compile")

}
