% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify_txt.R
\name{minify_txt}
\alias{minify_txt}
\title{Minifies text by removing extra white space}
\usage{
minify_txt(q)
}
\arguments{
\item{q}{text to be minified}
}
\value{
minified text
}
\description{
Minifies text by removing extra white space
}
\examples{
minify_txt(" This is   a lot of whitespace    ")
}
