% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsql_to_ducksql.R
\name{plsql_to_ducksql}
\alias{plsql_to_ducksql}
\title{Transform ORACLE Date function to be compatible with DUCKDB}
\usage{
plsql_to_ducksql(q)
}
\arguments{
\item{q}{Oracle SQL code}
}
\value{
DuckDb compatible SQL
}
\description{
Transform ORACLE Date function to be compatible with DUCKDB
}
\examples{
plsql_to_ducksql("sysdate - 730")
}
